--- streamAPIandLambdaTask1/Person.java
@@ -1,36 +1,0 @@
-package streamAPIandLambdaTask1;
-
-public class Person {
-    int age;
-    String name;
-
-    public Person() {
-    }
-
-    public Person(int age, String name) {
-        this.age = age;
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-
-    @Override
-    public String toString() {
-        return "Name = " + getName() + " ::: " + "Age = " + getAge();
-    }
-}
